pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpipe"
version = "0.1.0"
description = "Multi-provider LLM CLI tools"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "click>=8.1",
  "rich>=13.7",
  "requests>=2.32",
  "chromadb>=0.5",
  "Pillow>=10.0",
  "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
mpipe = "mpipe.main:main"
mpask = "mpipe.bin.mpask:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["mpipe*"]
